rename model classes & code cleaning
