+++ api/controller/schemas.py
@@ -1,9 +1,20 @@
-from typing import List
+import sys
+from typing import Any, Collection, Dict, List, Optional, Union
 from pydantic import BaseModel, create_model
+
+from api.config import DEFAULT_TOP_K_READER, DEFAULT_TOP_K_RETRIEVER
+
 
 class ModelDetails(BaseModel):
     id_model: str
     type_model: str
+
+
+class Question(BaseModel):
+    questions: List[str]
+    filters: Optional[Dict[str, Optional[Union[str, List[str]]]]] = None
+    top_k_reader: int = DEFAULT_TOP_K_READER
+    top_k_retriever: int = DEFAULT_TOP_K_RETRIEVER
 
 
 class Answer(BaseModel):
@@ -17,6 +28,7 @@ class Answer(BaseModel):
     offset_end_int_doc: int
     document_id: str
     meta: create_model('Meta', name=(int, ...))
+
 
 class Answers(BaseModel):
     query: str
